# Minor read-side RabbitMQ perf optimization