--- a/.gitignore
+++ b/.gitignore
@@ -204,3 +204,4 @@ build/
 .nuget/
 .dotnet/
 .idea
+/src/Amqp/Akka.Streams.RabbitMq.Amqp.Benchmarks/BenchmarkDotNet.Artifacts